Fix sendgrid config
