+++ .env.example
@@ -3,3 +3,4 @@ DB_USER=root
 DB_PASSWORD=
 DB_PORT=3306
 SENDGRID_API_KEY='SENDGRID_API_KEY_SECRET'
+SENDGRID_FROM='user@email.com'
